+++ lib/src/exception/image_initialization_exception.dart
@@ -1,8 +1,0 @@
-/// Exception during image initialization.
-class ImageInitializationException implements Exception {
-  /// Create an ImageInitializationException by specifying a [message].
-  ImageInitializationException(this.message);
-
-  /// An error message.
-  final String message;
-}